PNG support!
git-svn-id: https://svn.code.sf.net/p/doomlegacy/svn/legacy/trunk@143 5da24444-3c17-0410-b9f4-a1bbb4243164

+++ include/i_system.h
@@ -18,6 +18,9 @@
 //
 //
 // $Log$
+// Revision 1.6  2004/08/18 14:35:20  smite-meister
+// PNG support!
+//
 // Revision 1.5  2004/08/12 18:30:29  smite-meister
 // cleaned startup
 //
@@ -29,39 +32,6 @@
 //
 // Revision 1.2  2002/12/03 10:23:46  smite-meister
 // Video system overhaul
-//
-// Revision 1.4  2002/08/24 11:57:28  vberghol
-// d_main.cpp is better
-//
-// Revision 1.3  2002/07/01 21:00:48  jpakkane
-// Fixed cr+lf to UNIX form.
-//
-// Revision 1.2  2002/06/28 10:57:24  vberghol
-// Version 133 Experimental!
-//
-// Revision 1.8  2001/04/27 13:32:14  bpereira
-// no message
-//
-// Revision 1.7  2001/02/24 13:35:20  bpereira
-// no message
-//
-// Revision 1.6  2000/10/21 08:43:29  bpereira
-// no message
-//
-// Revision 1.5  2000/10/02 18:25:45  bpereira
-// no message
-//
-// Revision 1.4  2000/04/25 19:49:46  metzgermeister
-// support for automatic wad search
-//
-// Revision 1.3  2000/04/16 18:38:07  bpereira
-// no message
-//
-// Revision 1.2  2000/02/27 00:42:10  hurdler
-// fix CR+LF problem
-//
-// Revision 1.1.1.1  2000/02/22 20:32:32  hurdler
-// Initial import into CVS (v1.29 pr3)
 //
 //-----------------------------------------------------------------------------
 
@@ -95,7 +65,7 @@ extern BOOL   bDX0300;
 void I_SysInit();
 
 // return free and total physical memory in the system
-ULONG I_GetFreeMem(ULONG *total);
+Uint32 I_GetFreeMem(Uint32 *total);
 
 
 /// returns current time in 1/TICRATE second tics
